remove unneeded version switches for ECH ClientHello fragmentation
These were only needed for Go 1.23.
